Refactored the spiders
